+++ packages/backend-proxy-middleware/src/base/proxy.ts
@@ -1,32 +1,31 @@
-import { HttpsProxyAgent } from 'https-proxy-agent';
-import type { ServerOptions } from 'http-proxy';
-import type { RequestHandler, Options } from 'http-proxy-middleware';
-import { createProxyMiddleware } from 'http-proxy-middleware';
-import i18n from 'i18next';
-import type { ClientRequest, IncomingMessage, ServerResponse } from 'http';
-import { ToolsLogger, type Logger, UI5ToolingTransport } from '@sap-ux/logger';
 import { AbapCloudEnvironment, createForAbapOnCloud } from '@sap-ux/axios-extension';
-import {
-    isAppStudio,
-    getDestinationUrlForAppStudio,
-    getCredentialsForDestinationService,
-    listDestinations,
-    isFullUrlDestination,
-    BAS_DEST_INSTANCE_CRED_HEADER
-} from '@sap-ux/btp-utils';
 import type { ServiceInfo } from '@sap-ux/btp-utils';
-import type { BackendConfig, DestinationBackendConfig, LocalBackendConfig } from './types';
+import {
+    BAS_DEST_INSTANCE_CRED_HEADER,
+    getCredentialsForDestinationService,
+    getDestinationUrlForAppStudio,
+    isAppStudio,
+    isFullUrlDestination,
+    listDestinations
+} from '@sap-ux/btp-utils';
+import { ToolsLogger, UI5ToolingTransport, type Logger } from '@sap-ux/logger';
+import type { ClientRequest, IncomingMessage, ServerResponse } from 'http';
+import type { ServerOptions } from 'http-proxy';
+import type { Options, RequestHandler } from 'http-proxy-middleware';
+import { createProxyMiddleware } from 'http-proxy-middleware';
+import { HttpsProxyAgent } from 'https-proxy-agent';
+import i18n from 'i18next';
 import translations from './i18n.json';
-
+import type { BackendConfig, DestinationBackendConfig, LocalBackendConfig } from './types';
 import type { ApiHubSettings, ApiHubSettingsKey, ApiHubSettingsService, BackendSystem } from '@sap-ux/store';
 import { AuthenticationType, BackendSystemKey, getService } from '@sap-ux/store';
-import { updateProxyEnv } from './config';
-import type { Url } from 'node:url';
-import { addOptionsForEmbeddedBSP } from '../ext/bsp';
-import { getProxyForUrl } from 'proxy-from-env';
-import type { Socket } from 'node:net';
-import type { Request } from 'express';
 import type connect from 'connect';
+import type { Request } from 'express';
+import type { Socket } from 'node:net';
+import type { Url } from 'node:url';
+import { getProxyForUrl } from 'proxy-from-env';
+import { addOptionsForEmbeddedBSP } from '../ext/bsp';
+import { updateProxyEnv } from './config';
 
 export type EnhancedIncomingMessage = (IncomingMessage & Pick<Request, 'originalUrl'>) | connect.IncomingMessage;
 
@@ -264,16 +263,16 @@ export async function enhanceConfigsForDestination(
  *
  * @param proxyOptions reference to a proxy options object that the function will enhance
  * @param system backend system information (most likely) read from the store
- * @param oAuthRequired if true then the OAuth flow is triggered to get cookies
+ * @param authType determines the authentication protocol to be used
  * @param tokenChangedCallback function to call if a new refreshToken is available
  */
 export async function enhanceConfigForSystem(
     proxyOptions: Options & { headers: object },
     system: BackendSystem | undefined,
-    oAuthRequired: boolean | undefined,
+    authType: AuthenticationType,
     tokenChangedCallback: (refreshToken?: string) => void
 ): Promise<void> {
-    if (oAuthRequired) {
+    if (authType === AuthenticationType.OAuth2RefreshToken) {
         if (system?.serviceKeys) {
             const provider = createForAbapOnCloud({
                 environment: AbapCloudEnvironment.Standalone,
@@ -284,9 +283,9 @@ export async function enhanceConfigForSystem(
             // sending a request to the backend to get token
             await provider.getAtoInfo();
         } else {
-            throw new Error('Cannot connect to ABAP Environment on BTP without service keys.');
+            throw new Error('Cannot connect to ABAP Environment on BTP using OAuth without service keys.');
         }
-    } else if (system?.authenticationType === AuthenticationType.ReentranceTicket) {
+    } else if (system && authType === AuthenticationType.ReentranceTicket) {
         const provider = createForAbapOnCloud({
             ignoreCertErrors: proxyOptions.secure === false,
             environment: AbapCloudEnvironment.EmbeddedSteampunk,
@@ -336,7 +335,6 @@ export async function generateProxyMiddlewareOptions(
         target: backend.url,
         pathRewrite: PathRewriters.getPathRewrite(backend, logger)
     };
-
     // overwrite url if running in AppStudio
     if (isAppStudio()) {
         const destBackend = backend as DestinationBackendConfig;
@@ -346,39 +344,7 @@ export async function generateProxyMiddlewareOptions(
             logger.info('Using destination: ' + destBackend.destination);
         }
     } else {
-        const localBackend = backend as LocalBackendConfig;
-        // check if system credentials are stored in the store
-        try {
-            const systemStore = await getService<BackendSystem, BackendSystemKey>({ logger, entityName: 'system' });
-            const system = (await systemStore.read(
-                new BackendSystemKey({ url: localBackend.url, client: localBackend.client })
-            )) ?? {
-                name: '<unknown>',
-                url: localBackend.url,
-                authenticationType: localBackend.authenticationType
-            };
-            await enhanceConfigForSystem(
-                proxyOptions,
-                system,
-                backend.scp,
-                (refreshToken?: string, accessToken?: string) => {
-                    if (refreshToken) {
-                        logger.info('Updating refresh token for: ' + localBackend.url);
-                        systemStore.write({ ...system, refreshToken }).catch((error) => logger.error(error));
-                    }
-
-                    if (accessToken) {
-                        logger.info('Setting access token');
-                        proxyOptions.headers['authorization'] = `bearer ${accessToken}`;
-                    } else {
-                        logger.warn('Setting of access token failed.');
-                    }
-                }
-            );
-        } catch (error) {
-            logger.warn('Accessing the credentials store failed.');
-            logger.debug(error as object);
-        }
+        await updateProxyConfigFromStore(backend, logger, proxyOptions);
     }
 
     if (!proxyOptions.auth && process.env.FIORI_TOOLS_USER && process.env.FIORI_TOOLS_PASSWORD) {
@@ -409,6 +375,55 @@ export async function generateProxyMiddlewareOptions(
 
     logger.info(`Backend proxy created for ${proxyOptions.target}`);
     return proxyOptions;
+}
+
+/**
+ * Determine the correct authentication configuration for connections from a non-BAS platform.
+ *
+ * @param backend the backend config loaded from the yaml config
+ * @param logger a logger instance
+ * @param proxyOptions additional proxy header, request and response settings
+ */
+async function updateProxyConfigFromStore(
+    backend: BackendConfig,
+    logger: ToolsLogger,
+    proxyOptions: Options<IncomingMessage, ServerResponse<IncomingMessage>> & { headers: object }
+) {
+    const localBackend = backend as LocalBackendConfig;
+    // check if system credentials are stored in the store
+    try {
+        const systemStore = await getService<BackendSystem, BackendSystemKey>({ logger, entityName: 'system' });
+        const system = (await systemStore.read(
+            new BackendSystemKey({ url: localBackend.url, client: localBackend.client })
+        )) ?? {
+            name: '<unknown>',
+            url: localBackend.url,
+            authenticationType: localBackend.authenticationType
+        };
+        // Auth type is determined from app config as we may have multiple stored systems with the same url/client using different auth types
+        await enhanceConfigForSystem(
+            proxyOptions,
+            system,
+            localBackend.authenticationType ??
+                (localBackend.scp ? AuthenticationType.OAuth2RefreshToken : AuthenticationType.Basic),
+            (refreshToken?: string, accessToken?: string) => {
+                if (refreshToken) {
+                    logger.info('Updating refresh token for: ' + localBackend.url);
+                    systemStore.write({ ...system, refreshToken }).catch((error) => logger.error(error));
+                }
+
+                if (accessToken) {
+                    logger.info('Setting access token');
+                    proxyOptions.headers['authorization'] = `bearer ${accessToken}`;
+                } else {
+                    logger.warn('Setting of access token failed.');
+                }
+            }
+        );
+    } catch (error) {
+        logger.warn('Accessing the credentials store failed.');
+        logger.debug(error as object);
+    }
 }
 
 /**