+++ packages/cds-odata-annotation-converter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/cds-odata-annotation-converter
 
+## 0.6.13
+
+### Patch Changes
+
+-   Updated dependencies [ccb5285]
+    -   @sap-ux/odata-vocabularies@0.4.17
+    -   @sap-ux/cds-annotation-parser@0.2.20
+
 ## 0.6.12
 
 ### Patch Changes