+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.4.57
+
+### Patch Changes
+
+-   Updated dependencies [06bc541]
+    -   @sap-ux/inquirer-common@0.8.4
+
+## 0.4.56
+
+### Patch Changes
+
+-   Updated dependencies [c5d7915]
+    -   @sap-ux/project-access@1.32.5
+    -   @sap-ux/inquirer-common@0.8.3
+
+## 0.4.55
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.8.2
+
+## 0.4.54
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.8.1
+
+## 0.4.53
+
+### Patch Changes
+
+-   Updated dependencies [4053369]
+    -   @sap-ux/inquirer-common@0.8.0
+
+## 0.4.52
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.51
+
 ## 0.4.51
 
 ### Patch Changes